Create README.md
This course is based on student feedback such as: We would like to learn how to make publication-quality plots as they are shown in high-profile genomics journals. In this course,
we invite the lead bioinformatics authors of high-quality genomics articles, ask them to bring the data from their published articles, and guide course participants in redoing some of the published figures. In the mornings we have lectures from the teachers. The afternoons are hands-on workshops where the course participants work in small groups of around five students. The students stay in the same group for the whole week and work on one paper. The bioinformatics lead author of that paper is the mentor for that group.